requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
django-elasticsearch-dsl==7.2.0
django-jsoneditor==0.1.6
django-simpleui==2021.8.17
djangorestframework==3.12.4
djangorestframework-api-key==2.0.0
drf-generators==0.5.0
drf-yasg==1.20.0
elasticsearch==7.14.0
elasticsearch-dsl==7.4.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.2
packaging==21.0
pandas==1.3.3
psycopg2==2.9.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
scipy==1.7.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
